refactor(swig): consolidate setup into one swig:create step + separate grants
Collapse the per-step provisioning surface so the invariant part of every deploy
is a single command, and the only remaining scripts are the deploy-specific
grants and funding.

- Add `pnpm swig:create`: registers the Ledger owner, creates the Swig, and mints
  a bounded delegate (token/System programs + a default 0.1 SOL cap) in one
  command (2 approvals). Merges the old swig:init + swig:add-delegate happy path
  and bakes in the SOL cap, which the guide already called "not optional in
  practice" — removing the swap-succeeds-then-0xbbe footgun.
- Remove the all-in-one Path B (`swig:setup`/setup-swig-wallet.ts) and the
  now-redundant `swig:init`; delete the self-described-legacy
  `create-swig-wallet.ts` (`swig:provision`). Every provisioning path now mints a
  fresh, bounded delegate — reusing an existing wallet as the delegate is
  unsupported by design.
- Reframe `swig:add-delegate` as the rotate/add-another tool against an existing
  Swig; it too bakes in the SOL cap.
- lib.ts: add `resolveDelegateSolLimit` (default 0.1, logged, overridable) and
  `buildFreshDelegateRole`; drop the orphaned `provisionSwig`.
- Rewrite SETUP.md into a single 5-step flow (create → allow venues → cap tokens
  → fund → verify); update swig README.
