+++ package.json
@@ -33,9 +33,7 @@
     "test:scripts": "GATEWAY_TEST_MODE=dev jest --runInBand ./test-scripts/*.test.ts",
     "typecheck": "tsc --noEmit",
     "wallet:create": "npx ts-node scripts/create-wallet.ts",
-    "swig:setup": "npx ts-node -P scripts/tsconfig.json scripts/swig/setup-swig-wallet.ts",
-    "swig:provision": "npx ts-node -P scripts/tsconfig.json scripts/swig/create-swig-wallet.ts",
-    "swig:init": "npx ts-node -P scripts/tsconfig.json scripts/swig/init-swig.ts",
+    "swig:create": "npx ts-node -P scripts/tsconfig.json scripts/swig/create-swig.ts",
     "swig:add-delegate": "npx ts-node -P scripts/tsconfig.json scripts/swig/add-delegate.ts",
     "swig:allow-program": "npx ts-node -P scripts/tsconfig.json scripts/swig/allow-program.ts",
     "swig:add-token": "npx ts-node -P scripts/tsconfig.json scripts/swig/add-token.ts",
